requirements.txt
Django~=3.0.8
cachetools
dj_database_url
django-oauth-toolkit==1.3.2
django-cors-middleware
djangorestframework-gis~=0.15
djangorestframework~=3.11.0
hdx-python-api==2.1.5
mercantile~=0.10.0
psycopg2==2.8.5
python-openid==2.2.5
social-auth-app-django==4.0.0
social-auth-core==3.3.3
pytz
pyyaml==4.2b1
raven~=6.4.0
requests==2.20.0
requests_oauthlib==0.8.0
pyparsing==2.2.0
oauthlib==3.1.0
gunicorn==19.9.0
osmium~=2.15.2
dramatiq[redis,watch]~=1.7.0
boto3~=1.9.238 # needed for cloudwatch
rasterio~=1.0.25
osm-export-tool==0.0.24
rtree==0.9.1
